bot-mes-and-tool/requirements.txt
scikit-learn
python-whois
requests
bs4